+++ NEWS.md
@@ -1,7 +1,13 @@
 # ReSources development version
 
-# ReSources 22.09.1
+## ReSources 22.11.1
 
-## Enhancements
+### Bug fixes
+- fix update of list names for sources and concentration tables after names where changed in target
+tables, values are not reset anymore (#34)
+
+## ReSources 22.09.1
+
+### Enhancements
 - check if uploaded data contains empty tables
 - use shinyalerts in data uploads with different types of alarms: "error", "warning", "success"